cleanup: cast/type-safety and magic-number checkpoint across libslic3r + calibration dialog
- GCode.cpp / PerimeterGenerator.cpp / CutSurface.cpp: replace C-style casts with static_cast, fix signed/unsigned loop indices (reverse loops use the size_t t-- idiom), factor repeated expressions (cut_ratio, perimeter_width); GCodeOutputStream::write_format now uses a std::vector instead of malloc/free for oversized buffers
- Brim.cpp / PrintConfig.cpp: name magic numbers as constexpr constants (BP1002 style)
- BridgeDetector.cpp: reword TODOs into notes; a duplicated dead 'else if' with an identical condition was removed
- GCode.hpp: m_brim_done map key int -> size_t; MultiPoint.hpp: drop a redundant double() cast; Polyline.hpp: to_thick_polylines width parameter coordf_t -> coord_t + comment fix
- CalibrationPressureAdvDialog.cpp: first-layer-speed handling and border-position tweaks in the PA calibration generator

Working-tree checkpoint so the repo is clean; not separately re-verified beyond the builds run during the session.

+++ src/libslic3r/MultiPoint.hpp
@@ -468,7 +468,7 @@ public:
     // Remove exact duplicates, return true if any duplicate has been removed.
     bool remove_duplicate_points();
     virtual void douglas_peucker(coord_t tolerance = SCALED_EPSILON) {
-        auto it_end = Slic3r::douglas_peucker(points.begin(), this->points.end(), this->points.begin(), double(tolerance));
+        auto it_end = Slic3r::douglas_peucker(points.begin(), this->points.end(), this->points.begin(), tolerance);
         assert(it_end <= points.end());
         points.resize(std::distance(points.begin(), it_end));
     }
